Leverage priority from job details v2
